pulling successfully from firebase. Scrolling issue with cart resolved
