+++ src/IndividualGame.js
@@ -1,18 +1,20 @@
 //IndividualGame.js
-
-import { getDatabase, ref, onValue } from 'firebase/database';
 import firebase from './firebase';
+import { getDatabase, ref, onValue, push } from 'firebase/database';
 
 function IndividualGame(props){
 
      // console.log(props);
+     const database = getDatabase(firebase);
+     const dbRef = ref(database);
      
      const itemToAdd = () => {
           const gameTitle = props.name;
           const gamePrice = props.price;
           const key = props.infoToPass
           const image = props.image;
-          props.handleClick({"name":gameTitle, "price":gamePrice, "key":key, "image": image});
+          // props.handleClick({"name":gameTitle, "price":gamePrice, "key":key, "image": image});
+          push(dbRef, { "name": gameTitle, "price": gamePrice, "key": key, "image": image})
      }
