+++ src/firebase.js
@@ -21,4 +21,6 @@ const firebaseConfig = {
 
 
 // Initialize Firebase
-const app = initializeApp(firebaseConfig);
+const firebase = initializeApp(firebaseConfig);
+
+export default firebase;
